requirements.txt
# Gemini API视频识别客户端依赖包
google-genai>=0.7.0
tkinter-tooltip>=2.0.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# 飞书API集成依赖包
lark-oapi>=1.2.4  # 飞书官方Python SDK
requests>=2.31.0  # HTTP请求库（已包含在上面）
jsonschema>=4.0.0  # JSON数据验证
retrying>=1.3.4  # 重试机制

# 文件夹监控功能依赖包
watchdog>=3.0.0  # 文件系统监控库

# 阿里云OSS文件上传依赖包
oss2>=2.18.0  # 阿里云对象存储SDK